pyproject.toml
[project]
name = "vaultctl"
version = "0.1.0"
description = "Generalized Ansible Vault management CLI"
readme = "README.md"
authors = [
    { name = "cdds-ab", email = "info@cdds.de" }
]
requires-python = ">=3.13"
license = { text = "MIT" }
keywords = ["ansible", "vault", "secrets", "cli", "credentials"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Environment :: Console",
]

dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "bandit>=1.8.0",
    "commitizen>=4.9.1",
    "pre-commit>=4.0.1",
    "python-semantic-release>=10.4.1",
    "radon>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/cdds-ab/vaultctl"
Documentation = "https://github.com/cdds-ab/vaultctl#readme"
Repository = "https://github.com/cdds-ab/vaultctl"
Issues = "https://github.com/cdds-ab/vaultctl/issues"
Changelog = "https://github.com/cdds-ab/vaultctl/releases"

[project.scripts]
vaultctl = "vaultctl.cli:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

# Ruff configuration
[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src/vaultctl", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM117", # nested with statements - acceptable in tests for clarity
    "TC001",  # type-checking imports - runtime imports are simpler
    "TC003",  # type-checking imports for stdlib - runtime imports are simpler
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",    # unused arguments are common in tests
    "S101",   # assert statements are OK in tests
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

# MyPy configuration
[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
warn_return_any = false

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
    "--cov=vaultctl",
    "--cov-report=term-missing",
    "--cov-fail-under=70",
]

[tool.coverage.run]
source = ["src/vaultctl"]
omit = [
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

# Bandit configuration
[tool.bandit]
exclude_dirs = ["tests", ".venv", "scripts"]
skips = ["B101", "B603", "B607"]  # assert_used, subprocess_without_shell_equals_true, start_process_with_partial_path

# Commitizen configuration
[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
version_files = [
    "pyproject.toml:version",
    "src/vaultctl/__init__.py:__version__",
]
tag_format = "v$version"
update_changelog_on_bump = true
major_version_zero = true

# Semantic Release configuration
[tool.semantic_release]
version_variables = ["src/vaultctl/__init__.py:__version__"]
version_toml = ["pyproject.toml:project.version"]
branch = "master"
upload_to_pypi = false
upload_to_release = true
hvcs = "github"
commit_message = "chore(release): bump version to {version}"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "feat",
    "fix",
    "perf",
    "style",
    "refactor",
    "test",
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "style", "refactor"]

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "^chore",
    "^ci",
    "^build",
    "^style",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.4",
    "mypy>=1.14.0",
    "bandit>=1.8.0",
    "commitizen>=4.9.1",
    "pre-commit>=4.0.1",
    "python-semantic-release>=10.4.1",
    "radon>=6.0.1",
    "types-pyyaml>=6.0.0",
]
